requirements.txt
python-telegram-bot[job-queue]>=20.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
openai>=1.0.0
pytz>=2023.3
pytz>=2023.3
